refactor!: move template out of internal
