--- src/internal/template/testdata/invalid/input.yaml
@@ -1,1 +1,0 @@
-name: {{ .Values.missing
