+++ src/pkg/packager/template/refs.go
