+++ src/pkg/packager/template/refs_test.go
