+++ src/pkg/packager/template/testdata/configmap/expected.yaml
