+++ src/pkg/packager/template/testdata/configmap/input.yaml
