+++ src/pkg/packager/template/testdata/constants/input.yaml
