+++ src/pkg/packager/template/testdata/deployment/expected.yaml
