+++ src/pkg/packager/template/testdata/deployment/input.yaml
