+++ src/pkg/packager/template/testdata/invalid/input.yaml
@@ -1,0 +1,1 @@
+name: {{ .Values.missing
