+++ src/pkg/packager/template/testdata/sprig-basic/expected.yaml
