+++ src/pkg/packager/template/testdata/sprig-basic/input.yaml
