+++ src/pkg/packager/template/testdata/sprig-functions/expected.yaml
