+++ src/pkg/packager/template/testdata/sprig-functions/input.yaml
